third_party/blink/web_tests/external/wpt/css/css-align/content-distribution/place-content-shorthand-002-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Checking place-content: last baseline normal
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline left
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline right
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline start
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline end
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline center
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline flex-start
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline flex-end
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline stretch
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline space-around
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline space-between
  assert_equals: align-content expanded value expected "last baseline" but got ""
[FAIL] Checking place-content: last baseline space-evenly
  assert_equals: align-content expanded value expected "last baseline" but got ""
Harness: the test ran to completion.
